+++ PHPCompatibility/Util/Tests/Helpers/ResolveHelper/GetFQClassNameFromDoubleColonTokenUnitTest.inc
